refactor(offline): delete appInstanceResource before file
create listener file for eventListeners
