+++ src/actions/file.js
@@ -1,60 +1,16 @@
-import _ from 'lodash';
 import { DEFAULT_DELETE_REQUEST } from '../config/api';
 import { getApiContext, isErrorResponse, postMessage } from './common';
 import {
   DELETE_FILE_SUCCEEDED,
   DELETE_FILE_FAILED,
-  POST_FILE_FAILED,
-  POST_FILE_SUCCEEDED,
   DELETE_FILE,
 } from '../types';
-import {
-  postAppInstanceResource,
-  deleteAppInstanceResource,
-} from './appInstanceResources';
-import { showWarningToast } from '../utils/toasts';
-import {
-  FILE_UPLOAD_FAILED_MESSAGE,
-  FILE_DELETE_FAILED_MESSAGE,
-} from '../constants/messages';
 import { FILE } from '../config/appInstanceResourceTypes';
 
-const receiveFile = dispatch => event => {
-  const { data } = event;
-  try {
-    const message = JSON.parse(data);
-
-    const { type, payload } = message;
-
-    switch (type) {
-      case POST_FILE_SUCCEEDED:
-        return dispatch(postAppInstanceResource(payload));
-      case POST_FILE_FAILED: {
-        // the error message may be passed in payload
-        const errorMessage = _.isString(payload)
-          ? payload
-          : FILE_UPLOAD_FAILED_MESSAGE;
-        return showWarningToast(errorMessage);
-      }
-      case DELETE_FILE_SUCCEEDED:
-        return dispatch(deleteAppInstanceResource(payload));
-      case DELETE_FILE_FAILED: {
-        // the error message may be passed in payload
-        const errorMessage = _.isString(payload)
-          ? payload
-          : FILE_DELETE_FAILED_MESSAGE;
-        return showWarningToast(errorMessage);
-      }
-      default:
-        return false;
-    }
-  } catch (err) {
-    console.error(err);
-    return false;
-  }
-};
-
-const deleteFile = async ({ id, uri }) => async (dispatch, getState) => {
+const deleteFile = async ({ id, data: { uri } }) => async (
+  dispatch,
+  getState,
+) => {
   try {
     const {
       standalone,
@@ -91,8 +47,6 @@ const deleteFile = async ({ id, uri }) => async (dispatch, getState) => {
     // throws if it is an error
     await isErrorResponse(response);
 
-    dispatch(deleteAppInstanceResource({ id }));
-
     return dispatch({
       type: DELETE_FILE_SUCCEEDED,
     });
@@ -103,4 +57,5 @@ const deleteFile = async ({ id, uri }) => async (dispatch, getState) => {
   }
 };
 
-export { receiveFile, deleteFile };
+// eslint-disable-next-line import/prefer-default-export
+export { deleteFile };
