+++ src/actions/index.js
@@ -3,3 +3,4 @@ export * from './appInstanceResources';
 export * from './appInstance';
 export * from './users';
 export * from './layout';
+export * from './file';
